Second commit

+++ routers/todos.py
@@ -5,7 +5,7 @@ from database import SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
 from starlette import status
-
+from .auth import get_current_user
 
 router=APIRouter()
 
@@ -19,6 +19,8 @@ def get_db():
         db.close()
 
 dp_dependency = Annotated[Session, Depends(get_db) ]
+user_dependency = Annotated[dict,Depends(get_current_user)]
+
 
 class TodoRequset(BaseModel):
     title: str = Field(min_length=3)
@@ -27,25 +29,33 @@ class TodoRequset(BaseModel):
     complete:bool
 
 @router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(db: dp_dependency ):
-    return db.query(Todos).all()
+async def read_all(user:user_dependency,db: dp_dependency ):
+    if(user is None):
+        raise HTTPException(status_code=401,detail="Auth failed")
+    return db.query(Todos).filter(Todos.owner_id==user.get('user_id')).all()
 
 @router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def read_todo(db:dp_dependency,todo_id:int = Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def read_todo(user:user_dependency,db:dp_dependency,todo_id:int = Path(gt=0)):
+    if(user is None):
+        raise HTTPException(status_code=401,detail="Auth failed")
+    todo_model=db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('user_id')).first()
     if(todo_model is not None):
         return todo_model
     raise HTTPException(status_code=404,detail="Todo not found")
 
 @router.post("/todo/",status_code=status.HTTP_201_CREATED)
-async def create_todo(db:dp_dependency,todo_request:TodoRequset):
-    todo_model=Todos(**todo_request.model_dump())
+async def create_todo(user:user_dependency,db:dp_dependency,todo_request:TodoRequset):
+    if(user is None):
+        raise HTTPException(status_code=401,detail="Auth failed")
+    todo_model=Todos(**todo_request.model_dump(),owner_id=user.get('user_id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:dp_dependency,todo_request:TodoRequset,todo_id :int = Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def update_todo(user:user_dependency,db:dp_dependency,todo_request:TodoRequset,todo_id :int = Path(gt=0)):
+    if(user is None):
+        raise HTTPException(status_code=401,detail="Auth failed")
+    todo_model=db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('user_id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo not found")
     todo_model.title=todo_request.title
@@ -56,9 +66,11 @@ async def update_todo(db:dp_dependency,todo_request:TodoRequset,todo_id :int = P
     db.commit()
 
 @router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_book(db:dp_dependency,todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id==todo_id).first()
+async def delete_book(user:user_dependency,db:dp_dependency,todo_id: int = Path(gt=0)):
+    if(user is None):
+        raise HTTPException(status_code=401,detail="Auth failed")
+    todo_model = db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('user_id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo not found")
-    db.query(Todos).filter(Todos.id==todo_id).delete()
+    db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('user_id')).delete()
     db.commit()
